banking interface largely supported
The banking interface is now largely supported with only a few of my
originally intended features remaining to be added. Bank transactions
can now be added (along with information about the banks and accounts)
and transactions can be linked as transfers. In fact, these internal
links can also be used to link bank transactions with credit card
transactions, such as when a credit card payyment is made.

The process of updating linked payments is still a bit clunky (so right
now only creating joint transactions is supported; they must be modified
and deleted separately). The autosuggestion feature for bank transaction
notes is also not yet ideal.