Use the description as the deduced credit transaction merchant
While the description field is almost never going to be the final
value of the credit transaction field (it is often formatted weirdly),
the merchant suggestion alone does not always capture the best summary
of the fields contents. This is because it relies on previously entered
values, and if the app cannot connect the current merchant field to the
correct suggestion (or any suggestion), then information will be lost.
From experience, this can be annoying to the user. Instead, the app will
now provide the current transaction description field as the deduction
while also providing a better suggestion beneath.